+++ EMEProToolKit/EMEProToolkitSrc/AsyncContacts.cs
@@ -39,53 +39,6 @@ namespace EMEProToolkit
         private string _filePathEme = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\U.S. EPA\\EME Toolkit\\EMEdb\\";
         private string _installPath = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location);
 
-        public Task CopyContentsAsync(string srcDir, string targDir)
-        {
-
-            return Task.Run(() =>
-            {
-                //ArcGIS.Desktop.Framework.Dialogs.MessageBox.Show("CopyContentsAsync - copy contents task running");
-                if (!Directory.Exists(targDir))
-                {
-                    //ArcGIS.Desktop.Framework.Dialogs.MessageBox.Show("CopyContentsAsync - creating target dir");
-
-                    Directory.CreateDirectory(targDir);
-                    //Trace.WriteLine("Created target directory: "+targDir);
-                };
-                //ArcGIS.Desktop.Framework.Dialogs.MessageBox.Show("CopyContentsAsync - copying contents");
-
-                foreach (var f in Directory.GetFiles(srcDir))
-                {
-                    string fname = Path.GetFileName(f);
-                    //Trace.WriteLine("Copying " + fname);
-                    string dest = Path.Combine(targDir, fname);
-                    File.Copy(f, dest, overwrite: true);
-                }
-            });
-        }
-        public void DirectionsDir()
-        {
-            if (!Directory.Exists(_filePathEsri))
-                {
-                Directory.CreateDirectory(_filePathEsri);
-                }
-        }
-
-        public async void USEPADirAsync()
-        {
-            //ArcGIS.Desktop.Framework.Dialogs.MessageBox.Show("USEPADirAsync - Check if Target US EPA dir: "+ _filePathEme);
-            LogOutput.Log("USEPADirAsync - Check if Target US EPA dir: " + _filePathEme);
-
-            if (!Directory.Exists(_filePathEme))
-            {
-                string src = _installPath + "\\EMEdb\\";
-                //ArcGIS.Desktop.Framework.Dialogs.MessageBox.Show("USEPADirAsync - Creating US EPA db dir and copying contents from : " + src);
-                LogOutput.Log("USEPADirAsync - Creating U.S. EPA dir at: " + _filePathEme);
-                LogOutput.Log("USEPADirAsync - Source Dir: " + src);
-                await CopyContentsAsync(srcDir: src, targDir: _filePathEme);
-            }
-        }
-
         public Task LoadContactsAsync(bool checksyncage)
         {
             return Task.Run(() =>
