+++ EMEProToolKit/EMEProToolkitSrc/EMEProToolkit.csproj
@@ -101,6 +101,7 @@
   <ItemGroup>
     <Compile Include="AsyncContacts.cs" />
     <Compile Include="Daml.cs" />
+    <Compile Include="EMEInstall.cs" />
     <Compile Include="EMEMenu.cs" />
     <Compile Include="Logger.cs" />
     <Compile Include="Pages\BooleanValueSync.xaml.cs">
@@ -1036,7 +1037,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestButton1.cs" />
     <Compile Include="ThumbnailUpdater.cs" />
     <Compile Include="Utils\LanguageConverter.cs" />
     <Compile Include="Utils\MetadataException.cs" />
@@ -1047,7 +1047,6 @@
     <Compile Include="Sections\Overview.cs" />
     <Compile Include="Sections\Resource.cs" />
     <Compile Include="MetadataToolkitModule.cs" />
-    <Compile Include="ToolkitMetadataEditor.cs" />
     <AddInContent Include="Config.daml" />
     <AddInContent Include="Images\AddInDesktop16.png" />
     <AddInContent Include="Images\AddInDesktop32.png" />
